backend/requirements.txt
# Core dependencies
pdf2image>=1.16.3
pdfplumber>=0.7.6
pytesseract>=0.3.10
camelot-py[cv]>=0.10.1
PyMuPDF>=1.21.1
tabula-py>=2.7.0
easyocr>=1.7.0
python-dotenv>=1.0.0

# Optional dependencies
langchain>=0.0.267
unstructured>=0.8.1
opencv-python-headless>=4.7.0.72
numpy>=1.24.3
matplotlib>=3.7.1
Pillow>=9.5.0
requests>=2.28.2
